Removed relatives for rules window
For lower resolutions the rules window was broken. I removed relatives values and make the window centered. Also changed the size. Now it should be readable for all resolutions, but users with high ones would have problems. I suggest to double the size of the gui when 720p+ resolutions are detected.