Fixed small horns gridlists bug
When you open the horns tab, you cannot select the very first row now. You need to randomly click around on the gridlist and then after 1min of frustrating clicking it might allow you to select the first row. This fixes that so that the first row is selectable from the first try.