Update sb_c.lua
Changed date format to use dots as separators. The valid date format for Europe is 19.09.2020 and the one for US is 09/19/2020. Right now it shows a mixture of both, using the EU format but with US separators and it's anyone's guess what date it is if both day and month fall between 1 and 12.