Allow race checkpoint to use custom blip size (#816)
* Allow race checkpoints to use a custom blip size

Allows the blips for the dummy checkpoints in the sky in S.' recently added any order maps to be hidden.

MAP CREATORS:
You can now change the blip size of your checkpoints by adding a "blipsize" attribute to the checkpoint's node, and setting its value to an integer.
This is recommended for maps where the checkpoints are far away and bigger blips are desired so it's easier to find on the F11 map.
Alternatively this is useful for checkpoints that are activated by script instead of hit normally. For example "Pack Mule" has 100 checkpoints used to maintain some sense of logic in the progress bar and player ranking, which manifested themselves in the form of a black up-pointing triangle in the middle of the Los Santos bay.
Setting the blipsize to 0 will make it completely invisible. For reference: The default checkpoint blipsize is 2. The size of player blips and the upcoming checkpoint blip is 1. The size of the chased checkpoint blip in "Pack Mule" is 4.
Keep in mind that this is a change that applies to this server and 123robot's server only. Until we work out a way to get this feature approved into the official race gamemode, it is recommended that you continue to keep your scripty checkpoints tidy so your maps will also show up well in other and older versions of the race gamemode.

---------

Co-authored-by: SpeedyFolf <SpeedyFolf@Gmail.com>